% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-lib.R
\docType{data}
\name{JsonPagedScraper.class}
\alias{JsonPagedScraper.class}
\title{Json Paged Scraper class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JsonPagedScraper.class
}
\description{
Json Paged Scraper class
}
\keyword{datasets}
